+++ packages/orq-rc/RELEASES.md
@@ -10789,3 +10789,13 @@ Based on:
 - [python v4.15.0-rc.32] packages/orq-rc
 ### Releases
 - [PyPI v4.15.0-rc.32] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.32 - packages/orq-rc
+
+## 2026-09-09 14:48:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.15.0-rc.33] packages/orq-rc
+### Releases
+- [PyPI v4.15.0-rc.33] https://pypi.org/project/orq-ai-sdk/4.15.0-rc.33 - packages/orq-rc